+++ app/src/main/java/org/simple/clinic/consent/onboarding/UiActions.kt
@@ -1,0 +1,5 @@
+package org.simple.clinic.consent.onboarding
+
+interface UiActions {
+  fun moveToRegistrationActivity()
+}