+++ app/src/main/java/org/simple/clinic/onboarding/OnboardingEffectHandler.kt
@@ -1,17 +1,13 @@
 package org.simple.clinic.onboarding
 
-import com.f2prateek.rx.preferences2.Preference
 import com.spotify.mobius.functions.Consumer
 import com.spotify.mobius.rx2.RxMobius
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import io.reactivex.ObservableTransformer
-import org.simple.clinic.main.TypedPreference
-import org.simple.clinic.main.TypedPreference.Type.OnboardingComplete
 
 class OnboardingEffectHandler @AssistedInject constructor(
-    @TypedPreference(OnboardingComplete) private val hasUserCompletedOnboarding: Preference<Boolean>,
     @Assisted private val viewEffectsConsumer: Consumer<OnboardingViewEffect>
 ) {
 
@@ -25,16 +21,7 @@ class OnboardingEffectHandler @AssistedInject constructor(
   fun build(): ObservableTransformer<OnboardingEffect, OnboardingEvent> {
     return RxMobius
         .subtypeEffectHandler<OnboardingEffect, OnboardingEvent>()
-        .addTransformer(CompleteOnboardingEffect::class.java, completeOnboardingTransformer())
         .addConsumer(OnboardingViewEffect::class.java, viewEffectsConsumer::accept)
         .build()
-  }
-
-  private fun completeOnboardingTransformer(): ObservableTransformer<CompleteOnboardingEffect, OnboardingEvent> {
-    return ObservableTransformer { completeOnboardingEffect ->
-      completeOnboardingEffect
-          .doOnNext { hasUserCompletedOnboarding.set(true) }
-          .map { OnboardingCompleted }
-    }
   }
 }