+++ app/src/main/java/org/simple/clinic/onboarding/OnboardingUi.kt
@@ -1,5 +1,5 @@
 package org.simple.clinic.onboarding
 
 interface OnboardingUi {
-  fun moveToRegistrationScreen()
+  fun openOnboardingConsentScreen()
 }