+++ app/src/main/java/org/simple/clinic/setup/SetupActivityComponent.kt
@@ -3,12 +3,13 @@ package org.simple.clinic.setup
 import androidx.appcompat.app.AppCompatActivity
 import dagger.BindsInstance
 import dagger.Subcomponent
+import org.simple.clinic.consent.onboarding.OnboardingConsentScreenFragment
 import org.simple.clinic.navigation.v2.Router
 import org.simple.clinic.onboarding.OnboardingScreen
 import org.simple.clinic.splash.SplashScreen
 
 @Subcomponent(modules = [SetupActivityModule::class])
-interface SetupActivityComponent : OnboardingScreen.Injector, SplashScreen.Injector {
+interface SetupActivityComponent : OnboardingScreen.Injector, SplashScreen.Injector, OnboardingConsentScreenFragment.Injector {
 
   fun inject(target: SetupActivity)