+++ app/src/test/java/org/simple/clinic/onboarding/OnboardingEffectHandlerTest.kt
@@ -1,21 +1,17 @@
 package org.simple.clinic.onboarding
 
-import com.f2prateek.rx.preferences2.Preference
-import org.mockito.kotlin.eq
+import org.junit.After
+import org.junit.Test
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.verifyNoMoreInteractions
-import org.junit.After
-import org.junit.Test
 import org.simple.clinic.mobius.EffectHandlerTestCase
 
 class OnboardingEffectHandlerTest {
 
   private val uiActions = mock<OnboardingUi>()
-  private val hasUserCompletedOnboarding = mock<Preference<Boolean>>()
 
   private val effectHandler = OnboardingEffectHandler(
-      hasUserCompletedOnboarding = hasUserCompletedOnboarding,
       viewEffectsConsumer = OnboardingViewEffectHandler(uiActions)::handle
   ).build()
   private val testCase = EffectHandlerTestCase(effectHandler)
@@ -26,22 +22,12 @@ class OnboardingEffectHandlerTest {
   }
 
   @Test
-  fun `when complete onboarding effect is received, then mark user has completed onboarding`() {
+  fun `when open onboarding consent screen effect is received, then open onboarding consent screen`() {
     // when
-    testCase.dispatch(CompleteOnboardingEffect)
+    testCase.dispatch(OpenOnboardingConsentScreen)
 
     // then
-    verify(hasUserCompletedOnboarding).set(eq(true))
-    testCase.assertOutgoingEvents(OnboardingCompleted)
-  }
-
-  @Test
-  fun `when move to registration effect is received, then move to registration screen`() {
-    // when
-    testCase.dispatch(MoveToRegistrationEffect)
-
-    // then
-    verify(uiActions).moveToRegistrationScreen()
+    verify(uiActions).openOnboardingConsentScreen()
     verifyNoMoreInteractions(uiActions)
 
     testCase.assertNoOutgoingEvents()